requirements.txt
streamlit
gradio
crewai
openai
psycopg2-binary
python-dotenv
chromadb
PyMuPDF
pytesseract
Pillow
ollama
mistralai
redis
